move skimmers common functions to skimsUtils.h
